+++ lib/dp_exchange/webull/fake.ex
@@ -299,6 +299,12 @@ defmodule DpExchange.Webull.Fake do
   def get_conversion(_id, _opts \\ []), do: DpExchange.Core.Venue.not_supported()
 
   @impl true
+  def convert(_from, _to, _amount, _opts \\ []), do: Venue.not_supported()
+
+  @impl true
+  def get_trade_volume(_credentials, _opts \\ []), do: Venue.not_supported()
+
+  @impl true
   def list_portfolios(_opts \\ []), do: DpExchange.Core.Venue.not_supported()
 
   @impl true